autotrader/requirements.txt
alpaca-py>=0.21.0
ccxt>=4.2.0
yfinance>=0.2.36
python-dotenv>=1.0.0
loguru>=0.7.0
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
vectorbt>=0.26.0
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0
httpx>=0.27.0
anthropic>=0.39.0
python-telegram-bot>=21.0
apscheduler>=3.10.4
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
